make copylink into a file component
