+++ components/ActivityDetail.js
@@ -16,6 +16,7 @@ import Modal from './Modal'
 
 import useStore from '../store'
 import { parseImgUrl, prettyBalance, timeAgo } from '../utils/common'
+import CopyLink from './CopyLink'
 
 export const descriptionMaker = (activity, token) => {
 	if (activity.type === 'marketUpdate') {
@@ -39,40 +40,6 @@ export const descriptionMaker = (activity, token) => {
 	}
 
 	return `${activity.from} transferred ${activity.quantity}pcs of ${token?.metadata.name} to ${activity.to}`
-}
-
-const CopyLink = ({ children, link, afterCopy }) => {
-	const [isComponentMounted, setIsComponentMounted] = useState(false)
-	const copyLinkRef = useRef()
-
-	useEffect(() => {
-		setIsComponentMounted(true)
-	}, [])
-
-	const _copyLink = () => {
-		const copyText = copyLinkRef.current
-		copyText.select()
-		copyText.setSelectionRange(0, 99999)
-		document.execCommand('copy')
-
-		afterCopy()
-	}
-
-	return (
-		<div onClick={(_) => _copyLink()}>
-			{isComponentMounted && (
-				<div
-					className="absolute z-0 opacity-0"
-					style={{
-						top: `-1000`,
-					}}
-				>
-					<input ref={copyLinkRef} readOnly type="text" value={link} />
-				</div>
-			)}
-			<div className="relative z-10">{children}</div>
-		</div>
-	)
 }
 
 const Activity = ({ activity }) => {
@@ -173,9 +140,9 @@ const Activity = ({ activity }) => {
 
 const ActivityDetail = ({ activity, token }) => {
 	const [showModal, setShowModal] = useState(null)
-  const [isCopied, setIsCopied] = useState(false)
-  
-  const shareLink = `${process.env.BASE_URL}/activity/${activity._id}`
+	const [isCopied, setIsCopied] = useState(false)
+
+	const shareLink = `${process.env.BASE_URL}/activity/${activity._id}`
 
 	const fetcher = async (key) => {
 		const resp = await axios.get(`${process.env.API_URL}/${key}`)
@@ -208,10 +175,7 @@ const ActivityDetail = ({ activity, token }) => {
 			{showModal === 'options' && (
 				<Modal close={(_) => setShowModal('')}>
 					<div className="max-w-sm w-full px-4 py-2 bg-gray-100 m-auto rounded-md">
-						<CopyLink
-							link={shareLink}
-							afterCopy={handleAfterCopy}
-						>
+						<CopyLink link={shareLink} afterCopy={handleAfterCopy}>
 							<div className="py-2 cursor-pointer flex items-center">
 								<svg
 									className="rounded-md"
